addded dependencies and cleared previous creations
