--- src/App.css
@@ -1,16 +1,0 @@
-/*Nova spull*/
-body {
-  font-family: var(--normal-font)
-}
-header {
-  position: relative;
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background: url('./assets/images/header-background.jpg') no-repeat center center fixed;
-  background-size: cover;
-  color: white;
-}
