+++ src/App.js
@@ -4,31 +4,21 @@ import React, {useState} from "react";
 
 ///////////////////////
 //// Internal
-import './App.css';
+import './App.scss';
 import RectangleButton from "./components/shared/buttons/RectangleButton/RectangleButton";
-import Form from "./components/layout/Form/Form";
+import Form from "./components/layout/forms/Form/Form";
 import Navigation from "./components/layout/nav/Navigation";
 import UserTable from "./components/layout/UserTable/UserTable";
 
 function App() {
   const [value, setValue] = useState(``);
+  const [formState, setFormState] = useState({})
 
 
   return (
       <>
         <Navigation/>
         <header>
-          <h1>
-            Fruit perfection
-          </h1>
-            <Form/>
-          <RectangleButton onClick={() => console.log('Jij wil shoppen')} buttonSize={'btn--large'}
-                           buttonStyle={"btn--special--outline"}>
-            register
-          </RectangleButton>
-          <h1>
-            Imput testfield
-          </h1>
             <UserTable/>
