+++ src/App.scss
@@ -1,0 +1,12 @@
+header {
+  position: relative;
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  //background: url('./assets/images/header-background.jpg') no-repeat center center fixed;
+  background-size: cover;
+  color: white;
+}
