+++ src/index.scss
@@ -1,0 +1,55 @@
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;500;700;800&family=Press+Start+2P&family=Roboto+Mono:wght@100;300;700&display=swap');
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+:root {
+  --primary: hsl(214, 77%, 65%);
+  --secondary: hsl(214, 100%, 76%);
+  --tertiary: hsl(214, 51%, 41%);
+  --quaternary: hsl(213, 70%, 79%);
+  --quinary: hsl(39, 64%, 38%);
+  --senary: hsl(39, 77%, 65%);
+
+  --pine-tree: hsla(100, 18%, 13%, 1);
+  --blue-sapphire: hsla(195, 67%, 30%, 1);
+  --verdigris: hsla(180, 63%, 42%, 1);
+  --orange-yellow-crayola: hsla(47, 87%, 66%, 1);
+  --sandy-brown: hsla(28, 83%, 61%, 1);
+  /*https://coolors.co/1f271b-19647e-28afb0-f4d35e-ee964b*/
+
+  --warning: hsl(214, 51%, 41%);
+  --succes: hsl(213, 70%, 79%);
+  --danger: hsl(39, 64%, 38%);
+
+  --lighgray: hsl(0, 0%, 69%);
+  --darkgray: hsl(0, 0%, 36%);
+
+  --black: hsl(0, 0%, 14%);
+  --white: hsl(0, 0%, 100%);
+
+  --tittle-font: 'Press Start 2P', cursive;
+  --normal-font: 'Open Sans', sans-serif;
+
+  --select-border: var(--lighgray);
+  --select-focus: blue;
+  --select-arrow: var(--select-border);
+
+}
+
+h1 {
+  font-family: var(--tittle-font);
+}
+
+h2,h3,h4,h5 {
+  font-family: var(--normal-font);
+  font-weight: 700;
+}
+
+body {
+  font-family: var(--normal-font);
+  font-weight: 300;
+}
